# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  /** Price in BRL */
  price: number;
  imageUrl: string;
  category: string;
}

export interface OrderItemInput {
  menuItemId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItemInput[];
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 3 */
  customerEmail: string;
  /**
     * Optional CPF, improves PIX payment approval rates
     * @nullable
     */
  customerDocument?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  price: number;
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  awaiting_payment: 'awaiting_payment',
  approved: 'approved',
  rejected: 'rejected',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export interface Order {
  id: number;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  /**
     * PIX "copia e cola" payload
     * @nullable
     */
  pixQrCode: string | null;
  /**
     * Base64-encoded PNG of the PIX QR code
     * @nullable
     */
  pixQrCodeBase64: string | null;
  /** @nullable */
  pixExpiresAt: string | null;
}