# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  /** Price in BRL */
  price: number;
  imageUrl: string;
  category: string;
}