# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  /**
     * PIX "copia e cola" payload
     * @nullable
     */
  pixQrCode: string | null;
  /**
     * Base64-encoded PNG of the PIX QR code
     * @nullable
     */
  pixQrCodeBase64: string | null;
  /** @nullable */
  pixExpiresAt: Date | null;
}