# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItemInput[];
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerPhone: string;
  /** @minLength 3 */
  customerEmail: string;
  /**
     * Optional CPF, improves PIX payment approval rates
     * @nullable
     */
  customerDocument?: string | null;
  /** @nullable */
  notes?: string | null;
}